backend/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
pandas
httpx
python-dotenv
google-generativeai
greenlet
pydantic
pydantic-settings
